Added editing feature
